refactor: move useFilter out of internal directory
